FST - Refactor to use multiple engines
