+++ audio-faust/src/dsp/basic_fm.rs
@@ -36,67 +36,7 @@ Compilation options: -a /usr/local/share/faust/rust/jack-float.rs -lang rust -ct
 #![allow(unused_mut)]
 #![allow(non_upper_case_globals)]
 
-type F32 = f32;
-type F64 = f64;
-
-#[derive(Copy, Clone)]
-pub struct ParamIndex(pub i32);
-
-pub struct Soundfile<'a,T> {
-    fBuffers: &'a&'a T,
-    fLength: &'a i32,
-    fSR: &'a i32,
-    fOffset: &'a i32,
-    fChannels: i32
-}
-
-pub trait FaustDsp {
-    type T;
-
-    fn new() -> Self where Self: Sized;
-    fn metadata(&self, m: &mut dyn Meta);
-    fn get_sample_rate(&self) -> i32;
-    fn get_num_inputs(&self) -> i32;
-    fn get_num_outputs(&self) -> i32;
-    fn class_init(sample_rate: i32) where Self: Sized;
-    fn instance_reset_params(&mut self);
-    fn instance_clear(&mut self);
-    fn instance_constants(&mut self, sample_rate: i32);
-    fn instance_init(&mut self, sample_rate: i32);
-    fn init(&mut self, sample_rate: i32);
-    fn build_user_interface(&self, ui_interface: &mut dyn UI<Self::T>);
-    fn build_user_interface_static(ui_interface: &mut dyn UI<Self::T>) where Self: Sized;
-    fn get_param(&self, param: ParamIndex) -> Option<Self::T>;
-    fn set_param(&mut self, param: ParamIndex, value: Self::T);
-    fn compute(&mut self, count: i32, inputs: &[&[Self::T]], outputs: &mut[&mut[Self::T]]);
-}
-
-pub trait Meta {
-    // -- metadata declarations
-    fn declare(&mut self, key: &str, value: &str);
-}
-
-pub trait UI<T> {
-    // -- widget's layouts
-    fn open_tab_box(&mut self, label: &str);
-    fn open_horizontal_box(&mut self, label: &str);
-    fn open_vertical_box(&mut self, label: &str);
-    fn close_box(&mut self);
-
-    // -- active widgets
-    fn add_button(&mut self, label: &str, param: ParamIndex);
-    fn add_check_button(&mut self, label: &str, param: ParamIndex);
-    fn add_vertical_slider(&mut self, label: &str, param: ParamIndex, init: T, min: T, max: T, step: T);
-    fn add_horizontal_slider(&mut self, label: &str, param: ParamIndex, init: T, min: T, max: T, step: T);
-    fn add_num_entry(&mut self, label: &str, param: ParamIndex, init: T, min: T, max: T, step: T);
-
-    // -- passive widgets
-    fn add_horizontal_bargraph(&mut self, label: &str, param: ParamIndex, min: T, max: T);
-    fn add_vertical_bargraph(&mut self, label: &str, param: ParamIndex, min: T, max: T);
-
-    // -- metadata declarations
-    fn declare(&mut self, param: Option<ParamIndex>, key: &str, value: &str);
-}
+use crate::dsp::*;
 
 #[cfg_attr(feature = "default-boxed", derive(default_boxed::DefaultBoxed))]
 #[repr(C)]
